requirements.txt
openai>=1.0.0
langchain>=0.1.0
faiss-cpu>=1.7.3
tqdm
python-dotenv
